+++ src/scene/triangle.cu
@@ -3,8 +3,8 @@
 namespace CGL { namespace SceneObjects {
 
 DEVICE bool test_intersect(Ray &r, const Vector3D &p1,
-                  const Vector3D &p2, const Vector3D &p3, double &t, double &u,
-                  double &v) {
+                  const Vector3D &p2, const Vector3D &p3, float &t, float &u,
+                  float &v) {
   Vector3D e1 = p2 - p1, e2 = p3 - p1;
   Vector3D normal = cross(e1, e2);
   // early termination
@@ -13,7 +13,7 @@ DEVICE bool test_intersect(Ray &r, const Vector3D &p1,
   }
 
   Vector3D s = r.o - p1, s1 = cross(r.d, e2), s2 = cross(s, e1);
-  double rse1 = 1 / (dot(s1, e1));
+  float rse1 = 1 / (dot(s1, e1));
 
   t = dot(s2, e2) * rse1, u = dot(s1, s) * rse1, v = dot(s2, r.d) * rse1;
   if (t < r.min_t || t > r.max_t || u < 0 || v < 0 || u + v > 1) {
@@ -25,7 +25,7 @@ DEVICE bool test_intersect(Ray &r, const Vector3D &p1,
 }
 
 DEVICE bool CudaTriangle::intersect(Ray &r, CudaIntersection *isect) {
-  double t,u,v;
+  float t,u,v;
   if (!test_intersect(r, p1, p2, p3, t, u, v)) {
     return false;
   }
